gitignore configured for VSCode and Gradle, gradle configuration files and folder redesign for Gradle/support for other languages
