+++ src/main/java/com/gmail/lairmartes/datastructure/stack/GenericStack.java
