+++ src/main/java/com/gmail/lairmartes/game/hanoitower/Disk.java
