+++ src/main/java/com/gmail/lairmartes/game/hanoitower/Pin.java
