+++ src/main/java/com/gmail/lairmartes/game/hanoitower/event/HanoiTowerListener.java
@@ -1,0 +1,9 @@
+package com.gmail.lairmartes.game.hanoitower.event;
+
+public interface HanoiTowerListener {
+
+    void hanoiTowerEvent(GameOverEvent event);
+    void fireDiskRemoved(PinEvent event);
+    void fireDiskAdded(PinEvent event);
+    void hanoiTowerEvent(GameStartEvent event);
+}
