+++ src/main/java/com/gmail/lairmartes/game/hanoitower/exception/InvalidMoveException.java
